+++ src/app/locales/en/utils/workflow-actions/unlink-atm-lambda-action.js
@@ -1,0 +1,8 @@
+export default {
+  title: 'Unlink',
+  successNotificationText: 'The lambda has been unlinked sucessfully.',
+  failureNotificationActionName: 'unlinking the lambda',
+  tip: {
+    cannotUnlinkAtmLambdaUsed: 'This lambda cannot be unlinked because it is used by at least one workflow schema in this inventory.',
+  },
+};